relocate header libraries to dedicated dir
